fix: added tables formatter to avoid code repeating
we've also added some new comments and created new utils classes
to have a cleaner code. The major changes have been made to
the ADLdetectionservice and to the RegisterActivitiesService.

refs: #MIOT-61
